requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argcomplete=1.12.3=pyhd8ed1ab_2
beautifulsoup4=4.10.0=pyha770c72_0
blas=1.0=openblas
bottleneck=1.3.2=py39hdd57654_1
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py39h06a4308_0
cffi=1.14.6=py39h400218f_0
chardet=4.0.0=py39hf3d152e_2
compressed-rtf=1.0.6=pyh9f0ad1d_0
cryptography=3.4.8=py39hbca0aa6_0
docx2txt=0.8=py_0
ebcdic=1.1.1=pyhd8ed1ab_0
ebooklib=0.17=py_0
et_xmlfile=1.1.0=py39h06a4308_0
freetype=2.10.4=h0708190_1
icu=58.2=hf484d3e_1000
imapclient=2.2.0=pyhd8ed1ab_0
importlib-metadata=4.8.2=py39hf3d152e_0
importlib_metadata=4.8.2=hd8ed1ab_0
jbig=2.1=h7f98852_2003
jpeg=9d=h36c2ea0_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=2.2.1=h9c3ff4c_0
libdeflate=1.7=h7f98852_5
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h516909a_0
libopenblas=0.3.13=h4367d64_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.3.0=hf544144_1
libwebp-base=1.2.0=h27cfd23_0
libxml2=2.9.12=h03d6c58_0
libxslt=1.1.34=hc22bd24_0
lxml=4.6.3=py39h9120a33_0
lz4-c=1.9.3=h9c3ff4c_1
msg-extractor=0.28.7=pyhd8ed1ab_1
ncurses=6.3=h7f8727e_2
numexpr=2.7.3=py39h4be448d_1
numpy=1.21.2=py39hd8d4704_0
numpy-base=1.21.2=py39h2b8c604_0
olefile=0.46=pyh9f0ad1d_1
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1l=h7f8727e_0
pandas=1.3.4=py39h8c16a72_0
pdfminer.six=20211012=pyhd8ed1ab_0
pillow=7.2.0=py39h6f3857e_2
pip=21.2.4=py39h06a4308_0
pycparser=2.21=pyhd8ed1ab_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-docx=0.8.11=pyhd8ed1ab_0
python-pptx=0.6.21=pyhd8ed1ab_0
python-tzdata=2021.5=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd3eb1b0_0
pytz-deprecation-shim=0.1.0.post0=py39hf3d152e_1
readline=8.1=h27cfd23_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_0
soupsieve=2.3.1=pyhd8ed1ab_0
speechrecognition=3.7.1=py39hf3d152e_3
sqlite=3.36.0=hc218d9a_0
textract=1.6.4=pyhd8ed1ab_0
tk=8.6.11=h1ccaba5_0
tzdata=2021e=hda174b7_0
tzlocal=4.1=py39hf3d152e_1
wheel=0.37.0=pyhd3eb1b0_1
xlrd=2.0.1=pyhd8ed1ab_3
xlsxwriter=3.0.2=pyhd8ed1ab_0
xz=5.2.5=h7b6447c_0
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.5.0=ha95c52a_0
